backend/pyproject.toml
[project]
name = "rag-pipeline"
version = "0.1.0"
description = "Website-based Knowledge Ingestion and Vectorization Pipeline for RAG System"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "cohere>=4.9.0",
    "qdrant-client>=1.8.0",
    "python-dotenv>=1.0.0",
    "lxml>=4.9.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0"
]
